+++ collector.go
@@ -18,12 +18,13 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"sync"
 	"time"
 
 	"github.com/galexrt/extended-ceph-exporter/collector"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/sirupsen/logrus"
+	"go.uber.org/zap"
 )
 
 var (
@@ -45,7 +46,7 @@ var (
 type ExtendedCephMetricsCollector struct {
 	ctx             context.Context
 	ctxTimeout      time.Duration
-	log             *logrus.Logger
+	logger          *zap.Logger
 	lastCollectTime time.Time
 	clients         map[string]*collector.Client
 	collectors      map[string]collector.Collector
@@ -57,11 +58,11 @@ type ExtendedCephMetricsCollector struct {
 	cacheMutex     sync.Mutex
 }
 
-func NewExtendedCephMetricsCollector(ctx context.Context, log *logrus.Logger, clients map[string]*collector.Client, collectors map[string]collector.Collector, ctxTimeout time.Duration, cachingEnabled bool, cacheDuration time.Duration) *ExtendedCephMetricsCollector {
+func NewExtendedCephMetricsCollector(ctx context.Context, logger *zap.Logger, clients map[string]*collector.Client, collectors map[string]collector.Collector, ctxTimeout time.Duration, cachingEnabled bool, cacheDuration time.Duration) *ExtendedCephMetricsCollector {
 	return &ExtendedCephMetricsCollector{
 		ctx:             ctx,
 		ctxTimeout:      ctxTimeout,
-		log:             log,
+		logger:          logger,
 		lastCollectTime: time.Unix(0, 0),
 		clients:         clients,
 		collectors:      collectors,
@@ -85,9 +86,9 @@ func (n *ExtendedCephMetricsCollector) Collect(outgoingCh chan<- prometheus.Metr
 
 		expiry := n.lastCollectTime.Add(n.cacheDuration)
 		if time.Now().Before(expiry) {
-			n.log.Debugf("Using cache. Now: %s, Expiry: %s, LastCollect: %s", time.Now().String(), expiry.String(), n.lastCollectTime.String())
+			n.logger.Debug(fmt.Sprintf("Using cache. Now: %s, Expiry: %s, LastCollect: %s", time.Now().String(), expiry.String(), n.lastCollectTime.String()))
 			for _, cachedMetric := range n.cache {
-				n.log.Debugf("Pushing cached metric %s to outgoingCh", cachedMetric.Desc().String())
+				n.logger.Debug(fmt.Sprintf("Pushing cached metric %s to outgoingCh", cachedMetric.Desc().String()))
 				outgoingCh <- cachedMetric
 			}
 			return
@@ -107,11 +108,11 @@ func (n *ExtendedCephMetricsCollector) Collect(outgoingCh chan<- prometheus.Metr
 		for metric := range metricsCh {
 			outgoingCh <- metric
 			if n.cachingEnabled {
-				n.log.Debugf("Appending metric %s to cache", metric.Desc().String())
+				n.logger.Debug(fmt.Sprintf("Appending metric %s to cache", metric.Desc().String()))
 				n.cache = append(n.cache, metric)
 			}
 		}
-		n.log.Debug("Finished pushing metrics from metricsCh to outgoingCh")
+		n.logger.Debug("Finished pushing metrics from metricsCh to outgoingCh")
 	}()
 
 	wgCollection := sync.WaitGroup{}
@@ -131,10 +132,10 @@ func (n *ExtendedCephMetricsCollector) Collect(outgoingCh chan<- prometheus.Metr
 				var success float64
 
 				if err != nil {
-					n.log.WithError(err).Errorf("%s collector failed after %fs", collName, duration.Seconds())
+					n.logger.Error(fmt.Sprintf("%s collector failed after %fs", collName, duration.Seconds()), zap.Error(err))
 					success = 0
 				} else {
-					n.log.Debugf("%s collector succeeded after %fs.", collName, duration.Seconds())
+					n.logger.Debug(fmt.Sprintf("%s collector succeeded after %fs.", collName, duration.Seconds()))
 					success = 1
 				}
 				metricsCh <- prometheus.MustNewConstMetric(scrapeDurationDesc, prometheus.GaugeValue, duration.Seconds(), collName, clientName)
@@ -143,16 +144,16 @@ func (n *ExtendedCephMetricsCollector) Collect(outgoingCh chan<- prometheus.Metr
 		}
 	}
 
-	n.log.Debug("Waiting for collectors")
+	n.logger.Debug("Waiting for collectors")
 	wgCollection.Wait()
-	n.log.Debug("Finished waiting for collectors")
+	n.logger.Debug("Finished waiting for collectors")
 
 	n.lastCollectTime = time.Now()
-	n.log.Debugf("Updated lastCollectTime to %s", n.lastCollectTime.String())
+	n.logger.Debug(fmt.Sprintf("Updated lastCollectTime to %s", n.lastCollectTime.String()))
 
 	close(metricsCh)
 
-	n.log.Debug("Waiting for outgoing Adapter")
+	n.logger.Debug("Waiting for outgoing Adapter")
 	wgOutgoing.Wait()
-	n.log.Debug("Finished waiting for outgoing Adapter")
+	n.logger.Debug("Finished waiting for outgoing Adapter")
 }